UX-614 Convert Button to ts
